re #459: Fixed code review comments. Simplified Tracker threading.

+++ PlusLib/src/Tracking/vtkFakeTracker.h
@@ -77,7 +77,7 @@ protected:
 
 protected:
   /*! Internal frame number cunting */
-  int Frame;
+  unsigned long Frame;
 
   /*! Internal transform used for simulating tool movements */
   vtkTransform *InternalTransform;
